Refactor Staff actions hpp=>h/cpp
