+++ src/openrct2-ui/windows/StaffFirePrompt.cpp
@@ -10,7 +10,7 @@
 #include <openrct2-ui/interface/Widget.h>
 #include <openrct2-ui/windows/Window.h>
 #include <openrct2/Game.h>
-#include <openrct2/actions/StaffFireAction.hpp>
+#include <openrct2/actions/StaffFireAction.h>
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/interface/Colour.h>
 #include <openrct2/localisation/Localisation.h>
