+++ src/openrct2-ui/windows/StaffList.cpp
@@ -15,8 +15,8 @@
 #include <openrct2/Context.h>
 #include <openrct2/Game.h>
 #include <openrct2/Input.h>
-#include <openrct2/actions/StaffFireAction.hpp>
-#include <openrct2/actions/StaffSetColourAction.hpp>
+#include <openrct2/actions/StaffFireAction.h>
+#include <openrct2/actions/StaffSetColourAction.h>
 #include <openrct2/config/Config.h>
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/localisation/Localisation.h>
