+++ src/openrct2/actions/GameActionRegistration.cpp
@@ -69,13 +69,13 @@
 #include "SmallSceneryPlaceAction.h"
 #include "SmallSceneryRemoveAction.h"
 #include "SmallScenerySetColourAction.h"
-#include "StaffFireAction.hpp"
-#include "StaffHireNewAction.hpp"
-#include "StaffSetColourAction.hpp"
-#include "StaffSetCostumeAction.hpp"
-#include "StaffSetNameAction.hpp"
-#include "StaffSetOrdersAction.hpp"
-#include "StaffSetPatrolAreaAction.hpp"
+#include "StaffFireAction.h"
+#include "StaffHireNewAction.h"
+#include "StaffSetColourAction.h"
+#include "StaffSetCostumeAction.h"
+#include "StaffSetNameAction.h"
+#include "StaffSetOrdersAction.h"
+#include "StaffSetPatrolAreaAction.h"
 #include "SurfaceSetStyleAction.hpp"
 #include "TileModifyAction.hpp"
 #include "TrackDesignAction.h"
