+++ src/openrct2/libopenrct2.vcxproj
@@ -114,13 +114,13 @@
     <ClInclude Include="actions\SmallSceneryPlaceAction.h" />
     <ClInclude Include="actions\SmallSceneryRemoveAction.h" />
     <ClInclude Include="actions\SmallScenerySetColourAction.h" />
-    <ClInclude Include="actions\StaffFireAction.hpp" />
-    <ClInclude Include="actions\StaffHireNewAction.hpp" />
-    <ClInclude Include="actions\StaffSetColourAction.hpp" />
-    <ClInclude Include="actions\StaffSetCostumeAction.hpp" />
-    <ClInclude Include="actions\StaffSetNameAction.hpp" />
-    <ClInclude Include="actions\StaffSetOrdersAction.hpp" />
-    <ClInclude Include="actions\StaffSetPatrolAreaAction.hpp" />
+    <ClInclude Include="actions\StaffFireAction.h" />
+    <ClInclude Include="actions\StaffHireNewAction.h" />
+    <ClInclude Include="actions\StaffSetColourAction.h" />
+    <ClInclude Include="actions\StaffSetCostumeAction.h" />
+    <ClInclude Include="actions\StaffSetNameAction.h" />
+    <ClInclude Include="actions\StaffSetOrdersAction.h" />
+    <ClInclude Include="actions\StaffSetPatrolAreaAction.h" />
     <ClInclude Include="actions\SurfaceSetStyleAction.hpp" />
     <ClInclude Include="actions\TileModifyAction.hpp" />
     <ClInclude Include="actions\TrackDesignAction.h" />
