+++ src/openrct2/peep/Staff.cpp
@@ -12,8 +12,8 @@
 #include "../Context.h"
 #include "../Game.h"
 #include "../Input.h"
-#include "../actions/StaffHireNewAction.hpp"
-#include "../actions/StaffSetOrdersAction.hpp"
+#include "../actions/StaffHireNewAction.h"
+#include "../actions/StaffSetOrdersAction.h"
 #include "../audio/audio.h"
 #include "../config/Config.h"
 #include "../interface/Viewport.h"
